+++ tests/Integration/Services/CRM/Deal/Service/BatchTest.php
@@ -90,6 +90,55 @@ class BatchTest extends TestCase
         self::assertEquals(count($deals), $cnt);
     }
 
+    /**
+     * @testdox Batch delete deals
+     * @covers  \Bitrix24\SDK\Services\CRM\Deal\Service\Batch::update()
+     * @throws \Bitrix24\SDK\Core\Exceptions\BaseException
+     * @throws \Exception
+     */
+    public function testBatchUpdate(): void
+    {
+        // add deals
+        $deals = [];
+        for ($i = 1; $i < 60; $i++) {
+            $deals[] = ['TITLE' => 'TITLE-' . $i];
+        }
+        $cnt = 0;
+        $dealId = [];
+        foreach ($this->dealService->batch->add($deals) as $item) {
+            $cnt++;
+            $dealId[] = $item->getId();
+        }
+        self::assertEquals(count($deals), $cnt);
+
+        // read deals and prepare update information
+        $dealsToUpdate = [];
+        $resultDeals = [];
+        foreach ($this->dealService->batch->list([], ['ID' => $dealId], ['ID', 'TITLE', 'OPPORTUNITY']) as $deal) {
+            $dealOpportunity = random_int(100, 10000);
+            $dealsToUpdate[$deal->ID] = [
+                'fields' => [
+                    'OPPORTUNITY' => $dealOpportunity,
+                ],
+                'params' => [],
+            ];
+            $resultDeals[$deal->ID] = $dealOpportunity;
+        }
+
+        // update deals
+        foreach ($this->dealService->batch->update($dealsToUpdate) as $dealUpdateResult) {
+            $this->assertTrue($dealUpdateResult->isSuccess());
+        }
+
+        // list deals
+        $updateResult = [];
+        foreach ($this->dealService->batch->list([], ['ID' => $dealId], ['ID', 'TITLE', 'OPPORTUNITY']) as $deal) {
+            $updateResult[$deal->ID] = $deal->OPPORTUNITY;
+        }
+
+        $this->assertEquals($resultDeals, $updateResult);
+    }
+
     public function setUp(): void
     {
         $this->dealService = Fabric::getServiceBuilder()->getCRMScope()->deal();